remove amazon affiliate links
The Pro Git links to Amazon use Scott's personal affiliate
id, which puts Git the non-profit project in a murky area.
This site is not owned or managed by the Software Freedom
Conservancy's member project Git, but it's de facto status
as the official site makes things awkward. Dropping the
affiliate links removes a potential IRS headache for
Git/SFC.

In the long run, these may point to a non-profit affiliate
id for the Git project itself, but that id is not yet
available.